added a real "read-me", also changed some code, i think .-.

+++ src/jarra/JarraEx1_3.java
@@ -2,8 +2,8 @@ package jarra;
 
 public class JarraEx1_3 {
 	public static void main(String args[]) {
-		int x=10;
-		int y=x;
+		int		x=10;
+		double	y=x;
 		System.out.print(y); //imprime y que é igual a x 
 	}
 }
